LFS-safe r.drain
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@38632 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.drain/tinf.h
@@ -94,7 +94,7 @@ double slope_d(void *, void *, double);
 struct band3
 {
     int ns;			/* samples per line */
-    int sz;			/* bytes per line */
+    off_t sz;			/* bytes per line */
     char *b[3];			/* pointers to start of each line */
 };
